+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/distribution/ddorder/movement/schedule/DDOrderMoveScheduleService.java
@@ -6,6 +6,12 @@ import de.metas.ad_reference.ADReferenceService;
 import de.metas.distribution.ddorder.DDOrderId;
 import de.metas.distribution.ddorder.DDOrderLineId;
 import de.metas.distribution.ddorder.lowlevel.DDOrderLowLevelDAO;
+import de.metas.distribution.ddorder.movement.schedule.commands.drop_to.DDOrderDropToCommand;
+import de.metas.distribution.ddorder.movement.schedule.commands.drop_to.DDOrderDropToRequest;
+import de.metas.distribution.ddorder.movement.schedule.commands.material_in_trasit_report.MaterialInTransitReportCommand;
+import de.metas.distribution.ddorder.movement.schedule.commands.pick_from.DDOrderPickFromCommand;
+import de.metas.distribution.ddorder.movement.schedule.commands.pick_from.DDOrderPickFromRequest;
+import de.metas.distribution.ddorder.movement.schedule.commands.unpick.DDOrderUnpickCommand;
 import de.metas.distribution.ddorder.movement.schedule.plan.DDOrderMovePlan;
 import de.metas.distribution.ddorder.movement.schedule.plan.DDOrderMovePlanCreateCommand;
 import de.metas.distribution.ddorder.movement.schedule.plan.DDOrderMovePlanCreateRequest;
