+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/distribution/ddorder/movement/schedule/commands/drop_to/DDOrderDropToCommand.java
@@ -1,9 +1,12 @@
-package de.metas.distribution.ddorder.movement.schedule;
+package de.metas.distribution.ddorder.movement.schedule.commands.drop_to;
 
 import com.google.common.collect.ImmutableList;
 import de.metas.common.util.time.SystemTime;
 import de.metas.distribution.ddorder.lowlevel.DDOrderLowLevelDAO;
 import de.metas.distribution.ddorder.movement.generate.DDOrderMovementHelper;
+import de.metas.distribution.ddorder.movement.schedule.DDOrderMoveSchedule;
+import de.metas.distribution.ddorder.movement.schedule.DDOrderMoveScheduleId;
+import de.metas.distribution.ddorder.movement.schedule.DDOrderMoveScheduleRepository;
 import de.metas.handlingunits.movement.generate.HUMovementGenerateRequest;
 import de.metas.handlingunits.movement.generate.HUMovementGenerator;
 import de.metas.handlingunits.movement.generate.HUMovementGeneratorResult;
@@ -21,7 +24,7 @@ import javax.annotation.Nullable;
 import java.time.Instant;
 import java.util.Set;
 
-class DDOrderDropToCommand
+public class DDOrderDropToCommand
 {
 	@NonNull private final ITrxManager trxManager = Services.get(ITrxManager.class);
 	@NonNull private final DDOrderLowLevelDAO ddOrderLowLevelDAO;
