+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/distribution/ddorder/movement/schedule/commands/pick_from/DDOrderPickFromRequest.java
@@ -1,5 +1,6 @@
-package de.metas.distribution.ddorder.movement.schedule;
+package de.metas.distribution.ddorder.movement.schedule.commands.pick_from;
 
+import de.metas.distribution.ddorder.movement.schedule.DDOrderMoveScheduleId;
 import de.metas.handlingunits.HuId;
 import lombok.Builder;
 import lombok.NonNull;
